real_main_02/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.4.26
click==8.1.8
cx_Oracle==8.3.0
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
macholib==1.16.3
numpy==2.2.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
wheel==0.45.1
